backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.7
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faster-whisper==0.10.0
google-generativeai>=0.8.0
reportlab==4.0.7
python-pptx==0.6.23